Adjust check for whether or not the assume script has been sourced (#438)
* Add fallback support for any POSIX-compatible shell (#389)

* Add fallback support for any POSIX-compatible shell

As a last resort before erroring, try to determine a 'config' or
'profile' file to use via a POSIX-specific test.

[POSIX states regarding the ENV variable](https://pubs.opengroup.org/onlinepubs/9699919799/):

> This variable, when and only when an interactive shell is invoked, shall be
> subjected to parameter expansion (see Parameter Expansion) by the shell, and
> the resulting value shall be used as a pathname of a file containing shell
> commands to execute in the current environment. The file need not be
> executable.

My main shell is typically [mksh](http://www.mirbsd.org/mksh.htm). Even
though it lacks many of the bells and whistles of bash and zsh, it is an
adequate and comfortable interactive shell. By making all the common
pieces POSIX-compliant, all the currently supported shells should
continue to work and any additional shell that is POSIX-compliant should work
as well.

This also requires some changes in the assume shim script. Even though
some parts are zsh-specific and warnings can be ignored, the whole thing
is sourced and so needs to be syntactically correct:

- Replace '<<<' with a simple heredoc.
- Replace the 'source' command with '.'
- Use parameter substitution for testing the existence of substrings
- Remove some other unnecessary bashisms
- Minor tweak, but replace uses of `echo` with `printf` since that tends
  to be safer and more explicit in formatting. A good summary of the
  difference is [here](https://www.etalabs.net/sh_tricks.html)

* change 'return' to 'exit'

---------

Co-authored-by: chrnorm <17420369+chrnorm@users.noreply.github.com>

* Adjust check for whether or not the assume script has been sourced

See
https://github.com/common-fate/granted/pull/400#issuecomment-1606117287
for details

---------

Co-authored-by: chrnorm <17420369+chrnorm@users.noreply.github.com>
Co-authored-by: JoshuaWilkes <14214200+JoshuaWilkes@users.noreply.github.com>