Allow access token to be exported to plain text on ~/.aws/sso/cache
This change would allow the access tokens to be exported to ~/.aws/sso/cache
when the user run `assume --export`. Allowing `awscli` to use said token
without the need to login again. Adjusted the keys for the file name as
the default used is `sso_session_name`, `sso_start_url` is used if the
session name is not present. Fix for issue #155.