Allow access token to be exported to plain text on ~/.aws/sso/cache (#518)
* Allow access token to be exported to plain text on ~/.aws/sso/cache

This change would allow the access tokens to be exported to ~/.aws/sso/cache
when the user run `assume --export`. Allowing `awscli` to use said token
without the need to login again. Adjusted the keys for the file name as
the default used is `sso_session_name`, `sso_start_url` is used if the
session name is not present. Fix for issue #155.

* ✨Add plain-text sso cache check for assume

Added plain-text sso cache check for assume for users that has
valid access tokens created via `aws sso login` to avoid logging
in again. If the user has a valid plain-text, it will be stored
on the cache to avoid further logins if the access tokens are
still valid.

* ✨ Seperate export sso token to a separate flag

Added `--export-sso-token` flag to separate the plain text sso token
export from the `--export` flag. Also changed the text so it matches
with the flag.