🐛 Fix SSO token lookup keys
The lookup key for `GetValidSSOToken` has been updated and the code
I pushed had the wrong lookup causing panic when trying to export sso
as it was finding a `nil` token. Updated `GetValidSSOToken` to use
the combination of `SSOStartURL` and `SSOSessionName` to find it properly.
Also updated `GetValidSSOTokenFromPlaintextCache` to only use
`SSOStartURL` as it's the valid value. Should move to use `SSOSessionName`
in the future. Fix for issue #563.