Fix export lookup values when exporting credentials (#578)
* Fix export lookup values when exporting credentials

When `session` is present on the config it should be where
we get the `start_url` and `region`. Most configs will not
have said config under `profile` if they have a session block.
Session block takes priority for the configuration as it's considered
as it's connecting profiles together.

Fix for #577

* Adjust the logic for override

Override the sso_start_url if it's empty and have a proper session
block.