feat: Use aws:username for IAM initiated federated console sessions.
The changes refactor the way federation token ID is used for AWS IAM credentials.
Instead of relying on the userID which was previously parsed, the code now uses the
userName which is more easily attributable to the IAM user name in the Cloudtrail
events list view.

In the Cloudtrail console's event history view, the IAM user name will now display
in the `user name` column. Previously, the `user id` would display (e.g. AIDA.....).

Signed-off-by: Matthew Hembree <47449406+matthewhembree@users.noreply.github.com>