fix: don't set AWS profile when role chaining (#841)
Fixes https://github.com/common-fate/granted/issues/555

This is a crude attempt to the fix the issue by overwritting the profile
name when the chain flag is passed.  This is set at the end, right
before returning the values to the shell, the only functionality
difference should be that that the `AWS_PROFILE` environment variable
will not be set when the `--chain` flag is passed.